Fix liblbfgs summary
Setting a value in the PKG_SUMMARY array has no effect.
Therefore, even if PKG_SUMMARY[0] and PKG_SUMMARY[1] are set,
the subpackage summaries are not set
and the summary of main package is used.

Using PKG_CONTENTS array is deprecated.

This commit uses `eval`
for setting the subpackage summaries and contents.